Make the module resident rather than keeping a ref to it.
Rygel was crashing on exit because of modules being unloaded while there
were still objects of classes provided by the module. I thought I had
solved the issue by keeping a ref to each module in the PluginLoader but
the problem with that approach is that it is not guaranteed the
PluginLoader will not be freed before all the objects of classes it
provides.

git-svn-id: svn+ssh://svn.gnome.org/svn/rygel/trunk@318 d8cb91d7-bff9-0310-92b9-80b65e4482b2